+++ src/graphics/LairExplorerGraphic.h
@@ -11,7 +11,7 @@ public:
     LairExplorerGraphic(LairExplorer* explorer);
 
     void draw() override;
-    void update(uint32_t dt) override;
+    void update(float dt) override;
 
 private:
     Color m_color;
